added intersection distance; TODO (#intersect + #sym diff)/(#attrib1 + #attrib2 + min(#attrib1, #attrib2))

+++ src/main/java/cluster/DistanceFunction.java
@@ -5,5 +5,5 @@ import preprocess.DataObject;
 
 @FunctionalInterface
 public interface DistanceFunction<T extends DataObject> {
-  Integer calculate(DendrogramNode<T> node1, DendrogramNode<T> node2);
+  Float calculate(DendrogramNode<T> node1, DendrogramNode<T> node2);
 }
